+++ base/cvd/cuttlefish/host/commands/cvd/fetch/BUILD.bazel
@@ -98,7 +98,7 @@ cc_library(
         "//cuttlefish/common/libs/utils:flag_parser",
         "//cuttlefish/common/libs/utils:result",
         "//cuttlefish/host/commands/cvd/cache",
-        "//cuttlefish/host/libs/web/cas:cas_downloader",
+        "//cuttlefish/host/libs/web/cas:cas_flags",
         "//cuttlefish/host/libs/web:android_build_api",
         "//cuttlefish/host/libs/web:android_build_string",
         "//cuttlefish/host/libs/web:chrome_os_build_string",
